% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn3_source_rscience_modules.R
\name{fn3_source_rscience_modules}
\alias{fn3_source_rscience_modules}
\title{Source Science Modules}
\usage{
fn3_source_rscience_modules(paths, pattern = "^module_.*\\\\.R$")
}
\description{
Source Science Modules
}
