% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app001_InitPage.R
\name{run_app001_InitPage}
\alias{run_app001_InitPage}
\title{Ejecutar la aplicación Shiny}
\usage{
run_app001_InitPage()
}
\value{
No retorna valor.
}
\description{
Esta función inicia la aplicación Shiny incluida en este paquete.
}
