% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app2026.R
\name{run_app2026}
\alias{run_app2026}
\title{Ejecutar la aplicación Shiny}
\usage{
run_app2026()
}
\value{
No retorna valor.
}
\description{
Esta función inicia la aplicación Shiny incluida en este paquete.
}
